one/agents/subagents/coding_agent/agent.py
from google.adk.agents import LlmAgent , Agent
from google.adk.code_executors import BuiltInCodeExecutor

root_agent = Agent(
    name="coding_agent",
    model="gemini-2.0-flash",
    description=(
        "Sandboxed code and shell executor. Validates code examples and commands "
        "used in READMEs and answers."
    ),
    instruction="""
You are the code validator and executor. Use the BuiltInCodeExecutor only for:
 - Syntax checks
 - Running non-networked, safe commands (lint, unit tests, small compile)
 - Verifying that example commands work in principle and reporting outputs or errors.

Constraints:
 - Do not access the network.
 - Do not write or reveal secrets.
 - Limit file writes to a transient workspace.
 - Kill long-running processes after a short timeout.
 - When execution is not possible, provide deterministic static checks (lint, typecheck) and explain why you could not run.
Output:
 - For successful runs return { "status": "ok", "stdout": "...", "stderr": "..." }
 - For failures return { "status": "error", "error": "...", "diagnosis": "..." }
"""
    ,
    tools=[BuiltInCodeExecutor]
)
