pyproject.toml
[project]
name = "pr2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.2.0",
    "dotenv>=0.9.9",
    "google-adk>=1.13.0",
    "ipython>=9.4.0",
    "jupyterlab>=4.4.6",
    "langchain>=0.3.27",
    "langchain-community>=0.3.28",
    "langchain-groq>=0.3.7",
    "langchain-openai>=0.3.32",
    "langgraph>=0.6.6",
    "wikipedia>=1.4.0",
]
